added Carter, moved sensor endpoints to SensorEndpoints.cs #1
