+++ src/Devices.API/Devices.API.csproj
@@ -17,10 +17,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Infrastructure\" />
-    </ItemGroup>
-
-    <ItemGroup>
+      <PackageReference Include="Carter" Version="8.2.1" />
       <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.7" />
       <PackageReference Include="MongoDB.Driver" Version="2.28.0" />
       <PackageReference Include="Swashbuckle.AspNetCore" Version="6.7.0" />
