+++ src/Devices.API/Devices.API.http
@@ -1,11 +1,6 @@
 @Devices.API_HostAddress = http://localhost:5021
 
-GET {{Devices.API_HostAddress}}/todos/
-Accept: application/json
-
-###
-
-GET {{Devices.API_HostAddress}}/todos/1
+GET {{Devices.API_HostAddress}}/api/sensors
 Accept: application/json
 
 ###
