--- src/Devices.API/Features/CreateSensor/Models/CreateSensorCommand.cs
@@ -1,3 +1,0 @@
-﻿namespace Devices.API.Features.CreateSensor.Models;
-
-internal sealed record CreateSensorCommand();
