+++ src/Devices.API/Features/Sensors/CreateSensor/Models/CreateSensorCommand.cs
@@ -1,0 +1,3 @@
+﻿namespace Devices.API.Features.Sensors.CreateSensor.Models;
+
+internal sealed record CreateSensorCommand();
