+++ src/Devices.API/Program.cs
@@ -1,8 +1,7 @@
 using System.Text.Json.Serialization;
+using Carter;
 using Devices.API.Core;
-using Devices.API.Features.CreateSensor;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.OpenApi;
+using Devices.API.Features.Sensors.CreateSensor;
 using Microsoft.OpenApi.Models;
 
 namespace Devices.API;
@@ -22,12 +21,13 @@ public class Program
         {
             options.SerializerOptions.TypeInfoResolverChain.Insert(0, AppJsonSerializerContext.Default);
         });
+        builder.Services.AddCarter();
+        
         builder.Services.Configure<DevicesDatabaseSettings>(
             builder.Configuration.GetSection("DevicesDatabase"));
         builder.Services.AddSingleton<ISensorRepository, SensorRepository>();
 
         var app = builder.Build();
-
         if (app.Environment.IsDevelopment())
         {
             app.UseSwagger(c =>
@@ -36,40 +36,12 @@ public class Program
             });
             app.UseSwaggerUI();
         }
-
-        var sampleTodos = new Todo[]
-        {
-            new(1, "Walk the dog"),
-            new(2, "Do the dishes", DateOnly.FromDateTime(DateTime.Now)),
-            new(3, "Do the laundry", DateOnly.FromDateTime(DateTime.Now.AddDays(1))),
-            new(4, "Clean the bathroom"),
-            new(5, "Clean the car", DateOnly.FromDateTime(DateTime.Now.AddDays(2)))
-        };
-
-        var todosApi = app.MapGroup("/todos");
-        todosApi.MapPost("/", async ([FromServices] ISensorRepository sensorRepository) =>
-        {
-            await sensorRepository.CreateAsync(new Sensor("Test Sensor Name"));
-        }).WithOpenApi();
         
-        todosApi.MapGet("/", async ([FromServices] ISensorRepository sensorRepository) =>
-        {
-            var sensors = await sensorRepository.GetAllAsync();
-            
-            return Results.Ok(sensors);
-        }).WithOpenApi();
-        // todosApi.MapGet("/{id}", (int id) =>
-        //     sampleTodos.FirstOrDefault(a => a.Id == id) is { } todo
-        //         ? Results.Ok(todo)
-        //         : Results.NotFound());
-
+        app.MapCarter();
         app.Run();
     }
 }
 
-public record Todo(int Id, string? Title, DateOnly? DueBy = null, bool IsComplete = false);
-
-[JsonSerializable(typeof(Todo[]))]
 [JsonSerializable(typeof(List<Sensor>))]
 internal partial class AppJsonSerializerContext : JsonSerializerContext
 {
